# openapi.yaml
openapi: 3.0.1
info:
  title: RipoPlan Vehicle Intelligence API
  description: |
    API for VIN decoding, vehicle profile retrieval, diagnostics scanning,
    repair plan generation, and follow-up repair chat.
  version: "2.0.0"
servers:
  - url: https://{{HOST}}/
paths:
  /vin/decode:
    post:
      summary: Decode VIN from manual value or uploaded image
      operationId: decodeVin
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                vin:
                  type: string
                  description: VIN provided manually.
                image:
                  type: string
                  format: binary
                  description: Image containing a VIN plate/barcode.
      responses:
        '200':
          description: Decoded VIN details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VinDecodeResponse'
  /vehicle/{vin}:
    get:
      summary: Load enriched vehicle profile
      operationId: getVehicleProfile
      parameters:
        - in: path
          name: vin
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vehicle profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleProfile'
  /diagnostics/run:
    post:
      summary: Run diagnostics for a VIN
      operationId: runDiagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vin]
              properties:
                vin:
                  type: string
      responses:
        '200':
          description: Diagnostics result with issues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticsResponse'
  /repair/plan:
    post:
      summary: Build repair plan from diagnostics
      operationId: buildRepairPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vin, issues]
              properties:
                vin:
                  type: string
                issues:
                  type: array
                  items:
                    $ref: '#/components/schemas/DiagnosticIssue'
      responses:
        '200':
          description: Ordered repair steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairPlanResponse'
  /chat:
    post:
      summary: Ask a follow-up repair question
      operationId: vehicleChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vin, message]
              properties:
                vin:
                  type: string
                message:
                  type: string
                diagnostics:
                  type: array
                  items:
                    $ref: '#/components/schemas/DiagnosticIssue'
      responses:
        '200':
          description: Assistant response
          content:
            application/json:
              schema:
                type: object
                properties:
                  vin:
                    type: string
                  reply:
                    type: string
components:
  schemas:
    VinDecodeResponse:
      type: object
      properties:
        vin:
          type: string
        wmi:
          type: string
        manufacturer:
          type: string
        country:
          type: string
        model_year_code:
          type: string
    VehicleProfile:
      type: object
      properties:
        vin:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        hero_image:
          type: string
        specs:
          type: object
          additionalProperties:
            type: string
    DiagnosticIssue:
      type: object
      properties:
        code:
          type: string
        summary:
          type: string
        severity:
          type: string
    DiagnosticsResponse:
      type: object
      properties:
        vin:
          type: string
        summary:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosticIssue'
    RepairPlanResponse:
      type: object
      properties:
        vin:
          type: string
        estimated_total:
          type: string
        steps:
          type: array
          items:
            type: object
            properties:
              issue_code:
                type: string
              action:
                type: string
              eta:
                type: string